Anonymize images on client machine and convert to DICOM.
The original image is not sent to the service. Instead we read it,
remove all metadata, and save it in DICOM format using a generic
file name (some people use file names to encode information and we
don't want that). This change also enables this front end to support
all file formats SimpleITK does even if the service itself doesn't.