Update the main workflow and the pre-commit settings.
Updated the workflow to use newer versions of actions and to run
pre-commit not as an action. Updated the pre-commit configuration
to use newer versions of black and flake8. Also, the flake8 repository
pointed to on gitlab was a mirror of the github repo, so now pointing
to the original.