Moving to use the SimpleITK Utilities package.
Replaced the local resize and sitk2qpixmap functions with those in the
utilities package. Also added some keyboard shortcuts and fixed an
issue with the loading of images from csv (using relative paths with
respect to the csv file location instead of absolute paths).

As part of the move we also switch to PySide6 from PySide2.