Changing the display of the larger image.
When an image was selected for detailed display, it would be displayed
at full resolution. Most often this did not fully fit on screen and it
was impossible to see the whole image. Now the whole image is
displayed in the window and the user can zoom in/out to view the finer
details.